[#72] Update deploy script with dependencies
* Add deploy.jar
* Remove unused jar
* Update bash script

+++ util/upload-apk/script/flow-releases.sh
@@ -13,6 +13,12 @@ UPLOAD_S3_PATH=/path/to/uploadS3jar
 APK_BUILD_DIR=/path/to/apk/build/directory
 SERVER_CONFIG_PATH=/path/to/akvo-flow-server-config/repo
 
+# Configuration
+S3_ACCESS_KEY=s3_access_key
+S3_SECRET_KEY=s3_secret_key
+GAE_USERNAME=username
+GAE_PASSWORD=password
+
 VERSION=$(sed -n '/android:versionName="/{;s///;s/".*$//;p;d;}' AndroidManifest.xml | tr -d ' ')
 
 rm -r tmp;
@@ -26,9 +32,15 @@ for i in $(cat tmp/instances.txt); do
     rm -r bin;
     rm -r gen;
     echo '=================================================='
-    echo 'generating apk version' $VERSION 'for instance' $i;
-    ant flow-release -Dsurvey.properties=$SERVER_CONFIG_PATH/$i/survey.properties > tmp/antout.txt; 
-    mkdir -p builds/$i/$VERSION;
-    mv bin/fieldsurvey-*.apk builds/$i/$VERSION/; 
-    java -jar $UPLOAD_S3_PATH/uploadS3.jar S3_ACCESS_KEY S3_SECRET_KEY $i $APK_BUILD_DIR/$i/$VERSION/fieldsurvey-$VERSION.apk
+    if [ -f $SERVER_CONFIG_PATH/$i/survey.properties ]
+    then
+        echo 'generating apk version' $VERSION 'for instance' $i;
+        ant flow-release -Dsurvey.properties=$SERVER_CONFIG_PATH/$i/survey.properties >> tmp/antout.txt;
+        mkdir -p builds/$i/$VERSION;
+        mv bin/fieldsurvey-*.apk builds/$i/$VERSION/; 
+        echo 'Deploying APK...'
+        java -jar $UPLOAD_S3_PATH/deploy.jar $S3_ACCESS_KEY $S3_SECRET_KEY $i $APK_BUILD_DIR/$i/$VERSION/fieldsurvey-$VERSION.apk $VERSION $GAE_USERNAME $GAE_PASSWORD
+    else
+        'Cannot find survey.properties file for instance' $i;
+    fi
 done;
